rename RandomCode.h file
